Add a script that visualises photoresistor reads
Added a python script that displays an image that resembles the game
screen based on the photoresitor measurements for each frame.
Required dependencies: `pip install pyserial opencv-python numpy`

The main code running on the ATmega328 was simplified, all of the
terminal pretty printing was moved to the script. Debug serial output
from the microprocessor is minimal.